requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.2
django-ninja==1.3.0
nodeenv==1.9.1
pydantic==2.9.2
pydantic_core==2.23.4
pylyzer==0.0.44
pyright==1.1.386
pytest==8.0.0
pytest-django==4.8.0
ruff==0.8.0
sqlparse==0.5.1
typing_extensions==4.12.2
